(bug): fixing adding the same item in a cart functionnality

+++ client/my-app/src/screens/Cart.tsx
@@ -9,6 +9,7 @@ interface Product {
   name: string;
   price: number;
   imageUrl: string;
+  stock: number;
 }
 
 interface CartProduct {
@@ -25,10 +26,16 @@ interface Cart {
   updatedAt: string;
 }
 
+interface Order {
+  id: number;
+  cartId: number;
+  status: string;
+}
+
 const CartPage: React.FC = () => {
   const [carts, setCarts] = useState<Cart[]>([]);
   const [products, setProducts] = useState<Product[]>([]);
-  const [showProducts, setShowProducts] = useState<boolean>(false); // Contrôle d'affichage des produits
+  const [showProducts, setShowProducts] = useState<boolean>(false);
   const [selectedCart, setSelectedCart] = useState<number | null>(null);
   const [errorMessage, setErrorMessage] = useState<string | null>(null);
 
@@ -36,67 +43,100 @@ const CartPage: React.FC = () => {
   const navigate = useNavigate();
 
   useEffect(() => {
-    const fetchCarts = async () => {
+    const fetchCartsAndOrders = async () => {
       try {
-        const response = await axios.get('https://localhost:3000/market_place/v1/shop/user/carts', {
+        const cartResponse = await axios.get('https://localhost:3000/market_place/v1/shop/user/carts', {
           headers: { Authorization: `Bearer ${token}` },
         });
-        setCarts(response.data);
+        const orderResponse = await axios.get('https://localhost:3000/market_place/v1/shop/user/orders', {
+          headers: { Authorization: `Bearer ${token}` },
+        });
+
+        let unpaidCartIds: number[] = [];
+
+        if (orderResponse.data && Array.isArray(orderResponse.data) && orderResponse.data.length > 0) {
+          const unpaidOrders = orderResponse.data.filter((order: Order) => order.status === 'UNPAID');
+          unpaidCartIds = unpaidOrders.map((order: Order) => order.cartId);
+        }
+
+        const filteredCarts = unpaidCartIds.length > 0 
+          ? cartResponse.data.filter((cart: Cart) => unpaidCartIds.includes(cart.id))
+          : cartResponse.data;
+
+        setCarts(filteredCarts);
       } catch (error) {
-        console.error('Error fetching carts:', error);
+        console.error('Error fetching carts and orders:', error);
+        setErrorMessage('Erreur lors de la récupération des paniers et commandes.');
       }
     };
 
-    fetchCarts();
+    fetchCartsAndOrders();
   }, [token]);
 
-  // Fonction pour récupérer la liste des produits
   const fetchProducts = async () => {
     try {
       const response = await axios.get('https://localhost:3000/market_place/v1/shop/products', {
         headers: { Authorization: `Bearer ${token}` },
       });
-      setProducts(response.data.products); // S'assurer que la structure des données est correcte
-      setShowProducts(true); // Afficher la liste des produits
+      setProducts(response.data.products);
+      setShowProducts(true);
     } catch (error) {
       console.error('Erreur lors de la récupération des produits :', error);
       setErrorMessage('Erreur lors de la récupération des produits.');
     }
   };
 
-  const handleAddProduct = async (cartId: number, productId: number) => {
+  const handleAddProduct = async (cartId: number, product: Product) => {
+    if (product.stock <= 0) {
+      setErrorMessage(`Le produit ${product.name} est en rupture de stock.`);
+      return;
+    }
+
     try {
       await axios.put(
-        `https://localhost:3000/market_place/v1/shop/user/cart/${cartId}/${productId}`,
+        `https://localhost:3000/market_place/v1/shop/user/cart/${cartId}/${product.id}`,
         {
-          products: [{ id: productId, quantity: 1 }],
+          products: [{ id: product.id, quantity: 1 }],
         },
         {
           headers: { Authorization: `Bearer ${token}` },
         }
       );
-  
-      // Mise à jour du panier après ajout
+
       const response = await axios.get('https://localhost:3000/market_place/v1/shop/user/carts', {
         headers: { Authorization: `Bearer ${token}` },
       });
       setCarts(response.data);
-      setShowProducts(false); // Masquer la liste des produits après ajout
-      setSelectedCart(null); // Réinitialiser la sélection du panier
+      setShowProducts(false);
+      setSelectedCart(null);
     } catch (error) {
       console.error('Error adding product to cart:', error);
       setErrorMessage("Erreur lors de l'ajout du produit au panier.");
     }
   };
 
-  const handleRemoveProduct = async (cartId: number, productId: number, quantity: number = 1) => {
+  const handleRemoveProduct = async (cartId: number, productId: number, currentQuantity: number) => {
     try {
-      await axios.delete(`https://localhost:3000/market_place/v1/shop/user/cart/Item/${cartId}/${productId}`, {
-        headers: { Authorization: `Bearer ${token}` },
-        data: { productId, quantity },
-      });
+      if (currentQuantity > 1) {
+        // Si la quantité est supérieure à 1, on la diminue
+        await axios.put(
+          `https://localhost:3000/market_place/v1/shop/user/cart/${cartId}/${productId}`,
+          {
+            products: [{ id: productId, quantity: -1 }],
+          },
+          {
+            headers: { Authorization: `Bearer ${token}` },
+          }
+        );
+      } else {
+        // Si la quantité est égale à 1, on supprime le produit du panier
+        await axios.delete(`https://localhost:3000/market_place/v1/shop/user/cart/Item/${cartId}/${productId}`, {
+          headers: { Authorization: `Bearer ${token}` },
+          data: { productId: productId, quantity: currentQuantity },
+        });
+      }
 
-      // Mise à jour du panier après suppression
+      // Mise à jour du panier après suppression ou diminution de la quantité
       const response = await axios.get('https://localhost:3000/market_place/v1/shop/user/carts', {
         headers: { Authorization: `Bearer ${token}` },
       });
@@ -114,11 +154,11 @@ const CartPage: React.FC = () => {
         { cartId: cartId.toString() }, 
         { headers: { Authorization: `Bearer ${token}` } }
       );
-      const newOrderId = response.data.id; 
-      navigate(`/order-details/${newOrderId}/${cartId}`); 
+      const newOrderId = response.data.id;
+      navigate(`/order-details/${newOrderId}/${cartId}`);
     } catch (error) {
       console.error('Error creating order:', error);
-      alert('veuillez confirmer votre compte avant de passer une commande');
+      alert('Veuillez confirmer votre compte avant de passer une commande');
     }
   };
 
@@ -127,7 +167,7 @@ const CartPage: React.FC = () => {
       <h1>Vos paniers</h1>
       {errorMessage && <p className="error-message">{errorMessage}</p>}
       {carts.length === 0 ? (
-        <p>Votre panier est vide.</p>
+        <p>Votre panier est vide ou payé.</p>
       ) : (
         <div className="panier-grid">
           {carts.map(cart => (
@@ -148,7 +188,7 @@ const CartPage: React.FC = () => {
                         <p>Quantité: {quantity}</p>
                       </div>
                       <div className="panier-actions">
-                        <button onClick={() => handleRemoveProduct(cart.id, product.id)}>
+                        <button onClick={() => handleRemoveProduct(cart.id, product.id, quantity)}>
                           <FaMinus /> Retirer
                         </button>
                         <button onClick={() => { setSelectedCart(cart.id); fetchProducts(); }}>
@@ -174,7 +214,6 @@ const CartPage: React.FC = () => {
         </div>
       )}
 
-      {/* Affichage de la liste des produits */}
       {showProducts && (
         <div className="product-list">
           <h2>Ajouter un produit</h2>
@@ -184,7 +223,11 @@ const CartPage: React.FC = () => {
                 <img src={product.imageUrl} alt={product.name} />
                 <h3>{product.name}</h3>
                 <p>Prix: {product.price} €</p>
-                <button onClick={() => handleAddProduct(selectedCart!, product.id)}>
+                <button 
+                  onClick={() => handleAddProduct(selectedCart!, product)}
+                  disabled={product.stock <= 0}
+                  className={product.stock <= 0 ? "disabled-add-to-cart" : ""}
+                >
                   Ajouter au panier
                 </button>
               </div>
